Renaming things because Pytest gets confused.
